moving items again so all ui is together
