+++ src/react/ui/components/graph/barGraph/barGraph.tsx
