+++ src/react/ui/components/graph/lineGraph/lineGraph.tsx
